sqs send order created and received event

+++ dto/create_order_event.go
@@ -1,6 +1,6 @@
 package dto
 
 type CreateOrderEvent struct {
-	OrderID    string `json:"order_id"`
-	TotalPrice int64  `json:"total_price"`
+	OrderID    string  `json:"order_id"`
+	TotalPrice float32 `json:"total_price"`
 }
